Vehicle Control

+++ Assets/Configs/CarConfig.cs
@@ -7,6 +7,6 @@ public class CarConfig : ScriptableObject
 {
     public int price;
     public int maxSpeed;
-    public int braking;
-    public int acceleration;
+    public float braking;
+    public float acceleration;
 }
